refactor: `Dashboard` -> `dashboard`
